+++ forge/src/main/java/cc/cassian/item_descriptions/client/forge/ItemDescriptionsForge.java
@@ -3,7 +3,7 @@ package cc.cassian.item_descriptions.client.forge;
 import cc.cassian.item_descriptions.client.ModClient;
 import cc.cassian.item_descriptions.client.config.ModConfig;
 import cc.cassian.item_descriptions.client.config.forge.ModConfigFactory;
-import cc.cassian.item_descriptions.client.helpers.GenericKeys;
+import cc.cassian.item_descriptions.client.helpers.TagHelpers;
 import net.minecraft.item.ItemStack;
 import net.minecraft.text.Text;
 import net.minecraftforge.client.ConfigScreenHandler;
@@ -44,7 +44,7 @@ public final class ItemDescriptionsForge {
             ItemStack stack = event.getItemStack();
             List<Text> tooltip;
             if (ModConfig.get().developer_showAllPotentialKeys) {
-                tooltip = GenericKeys.findAllPotentialKeys(stack);
+                tooltip = TagHelpers.findAllPotentialKeys(stack);
             }
             else {
                 tooltip = createTooltip(findItemLoreKey(stack), false);