+++ neoforge/src/main/java/cc/cassian/item_descriptions/client/helpers/neoforge/ModHelpersImpl.java
@@ -6,7 +6,7 @@ public class ModHelpersImpl {
     public static boolean clothConfigInstalled() {
         return ModList.get().isLoaded("cloth_config");
     }
-    public static boolean tooltipFixInstalled() {
-        return false;
+    public static boolean isLoaded(String mod) {
+        return ModList.get().isLoaded(mod);
     }
 }